added help support in MAVExplorer
can do:
  param help PARAMNAME
  param check
  param download
